revert alert channel
